move Courier to courier.rs
creating a module to make main.rs more readable and manageable
